+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 ##### Entities:
 Fixed Beehemoth not being saddleable and being controllable without being saddled.
 
+##### Items:
+Honey Compass now says the name of the block it is tracking when in block tracking mode.
+
 
 ### **(V.6.1.0 Changes) (1.19 Minecraft)**