+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+### **(V.6.1.1 Changes) (1.19 Minecraft)**
+
+##### Entities:
+Fixed Beehemoth not being saddleable and being controllable without being saddled.
+
+##### Items:
+Honey Compass now says the name of the block it is tracking when in block tracking mode.
+
+
 ### **(V.6.1.0 Changes) (1.19 Minecraft)**
 
 ##### Entities: